Moved excpetion.h away from global place. Also reduced clutter.
