+++ src/jsonlogic/exception.h
