+++ src/operations/array/in.cpp
@@ -5,8 +5,6 @@
 
 #include "nlohmann/json.hpp"
 
-#include "exception.h"
-
 using namespace nlohmann;
 
 namespace json_logic
