+++ src/operations/data/var.cpp
@@ -7,8 +7,6 @@
 
 #include "nlohmann/json.hpp"
 
-#include "exception.h"
-
 using namespace nlohmann;
 
 namespace json_logic
