+++ src/operations/string/cat.cpp
@@ -4,8 +4,6 @@
 
 #include "nlohmann/json.hpp"
 
-#include "exception.h"
-
 using namespace nlohmann;
 
 namespace json_logic
